requirements.txt
pillow==10.4.0
numpy==2.1.2
flask==3.0.3